Massive balance reworks to Enterprise Disposal
To-do list:

- Texture disposal machines
- Fix disposal machine GUIs - remove them maybe
- Figure out what to do with ordinary subatomic shredder
- Make heat spread more intuitively (as in not from large cold objects
to small hot objects)
- Add capacitors, perhaps? Then again, it's not all that hard to move 3K
RF/tick, especially with Thermal Expansion to help out. It's 
- Recipes for Enterprise Disposal
- Maybe more...

+++ src/main/java/me/planetguy/ore/content/lavawell/TileEntityThermometer.java
@@ -22,7 +22,7 @@ public class TileEntityThermometer extends TileEntity{
 		SimplePeripheral.registerSimplePeripheral(TileEntityThermometer.class);
 	}
 
-	private volatile int cachedTemperature;
+	private volatile float cachedTemperature;
 	
 	public void updateEntity(){
 		cachedTemperature=0;
